+++ src/Control/Effect/Lift.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE FlexibleContexts, MultiParamTypeClasses #-}
 module Control.Effect.Lift
 ( Lift(..)
-, lift
+, sendM
 , runM
 , LiftC(..)
 ) where
@@ -15,10 +15,11 @@ import Control.Effect.Lift.Internal
 runM :: Monad m => Eff (LiftC m) a -> m a
 runM = runLiftC . interpret
 
--- | Given a @Lift n@ constraint in a signature carried by @m@, use 'lift' to
--- promote arbitrary actions of type @n a@ to @m a@.
-lift :: (Member (Lift n) sig, Carrier sig m, Functor n, Applicative m) => n a -> m a
-lift = send . Lift . fmap pure
+-- | Given a @Lift n@ constraint in a signature carried by @m@, 'sendM'
+-- promotes arbitrary actions of type @n a@ to @m a@. It is spiritually
+-- similar to @lift@ from the @MonadTrans@ typeclass.
+sendM :: (Member (Lift n) sig, Carrier sig m, Functor n, Applicative m) => n a -> m a
+sendM = send . Lift . fmap pure
 
 newtype LiftC m a = LiftC { runLiftC :: m a }