+++ ChangeLog.md
@@ -1,3 +1,7 @@
+- Adds benchmarks of `WriterC`/`VoidC` wrapped with `Eff` against their unwrapped counterparts.
+- Adds `Functor`, `Applicative`, and `Monad` instances for `WriterC`.
+- Adds `Functor`, `Applicative`, and `Monad` instances for `VoidC`.
+- Fixes a space leak with `WriterC`.
 - Removes the `Functor` constraint on `asks`.
 
 # 0.1.2.1