+++ ChangeLog.md
@@ -6,6 +6,7 @@
 - Fixes a space leak with `WriterC`.
 - Removes the `Functor` constraint on `asks`.
 - Adds `bracketOnError` to `Resource`.
+- Adds `sendM` to `Lift`.
 
 # 0.1.2.1