+++ ChangeLog.md
@@ -4,7 +4,7 @@
 - Adds `Functor`, `Applicative`, and `Monad` instances for `WriterC`.
 - Adds `Functor`, `Applicative`, and `Monad` instances for `VoidC`.
 - Fixes a space leak with `WriterC`.
-- Removes the `Functor` constraint on `asks`.
+- Removes the `Functor` constraint on `asks` and `gets`.
 - Adds `bracketOnError`, `finally`, and `onException` to `Resource`.
 - Adds `sendM` to `Lift`.